add strict dependencies, add unit tests for services

+++ spring-09/homework-08/src/main/java/ru/otus/spring/service/AuthorServiceImpl.java
@@ -2,7 +2,9 @@ package ru.otus.spring.service;
 
 import org.springframework.stereotype.Component;
 import ru.otus.spring.domain.Author;
+import ru.otus.spring.domain.Book;
 import ru.otus.spring.repository.AuthorRepository;
+import ru.otus.spring.repository.BookRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -10,9 +12,14 @@ import java.util.Optional;
 @Component
 public class AuthorServiceImpl implements AuthorService {
     private final AuthorRepository authorRepository;
+    private final BookRepository bookRepository;
 
-    public AuthorServiceImpl(AuthorRepository authorRepository) {
+    public AuthorServiceImpl(
+            AuthorRepository authorRepository,
+            BookRepository bookRepository
+    ) {
         this.authorRepository = authorRepository;
+        this.bookRepository = bookRepository;
     }
 
     public Author add(String name) {
@@ -27,12 +34,26 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public Author save(Author author) {
-        return authorRepository.save(author);
+        Author updatedAuthor = authorRepository.save(author);
+        if (author.getId() != null) {
+            List<Book> books = bookRepository.findByAuthor(author);
+            books.forEach( book -> {
+                book.setAuthor(updatedAuthor);
+                bookRepository.save(book);
+            });
+        }
+        return updatedAuthor;
     }
 
     @Override
     public void delete(Author author) {
-        authorRepository.delete(author);
+        if (author.getId() != null) {
+            if (bookRepository.existsByAuthor(author)) {
+                throw new RuntimeException("Author can't be deleted due to the presence of books with the author = "
+                        + author.getName());
+            }
+            authorRepository.delete(author);
+        }
     }
 
     @Override
