+++ spring-09/homework-08/src/main/java/ru/otus/spring/service/GenreServiceImpl.java
@@ -1,7 +1,9 @@
 package ru.otus.spring.service;
 
 import org.springframework.stereotype.Component;
+import ru.otus.spring.domain.Book;
 import ru.otus.spring.domain.Genre;
+import ru.otus.spring.repository.BookRepository;
 import ru.otus.spring.repository.GenreRepository;
 
 import java.util.List;
@@ -10,9 +12,14 @@ import java.util.Optional;
 @Component
 public class GenreServiceImpl implements GenreService {
     private final GenreRepository genreRepository;
+    private final BookRepository bookRepository;
 
-    public GenreServiceImpl(GenreRepository genreRepository) {
+    public GenreServiceImpl(
+            GenreRepository genreRepository,
+            BookRepository bookRepository
+    ) {
         this.genreRepository = genreRepository;
+        this.bookRepository = bookRepository;
     }
 
     @Override
@@ -28,12 +35,26 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public Genre save(Genre genre) {
-        return genreRepository.save(genre);
+        Genre updatedGenre = genreRepository.save(genre);
+        if (genre.getId() != null) {
+            List<Book> books = bookRepository.findByGenre(genre);
+            books.forEach( book -> {
+                book.setGenre(updatedGenre);
+                bookRepository.save(book);
+            });
+        }
+        return updatedGenre;
     }
 
     @Override
     public void delete(Genre genre) {
-        genreRepository.delete(genre);
+        if (genre.getId() != null) {
+            if (bookRepository.existsByGenre(genre)) {
+                throw new RuntimeException("Genre can't be deleted due to the presence of books with the genre = "
+                        + genre.getName());
+            }
+            genreRepository.delete(genre);
+        }
     }
 
     @Override
